LPS-195086 Extract methods from DLSizeLimitManagedServiceFactory to DLSizeLimitConfigurationHelper and apply to usages
