+++ modules/apps/document-library/document-library-service/src/main/java/com/liferay/document/library/internal/configuration/persistence/listener/DLSizeLimitConfigurationModelListener.java
@@ -6,7 +6,6 @@
 package com.liferay.document.library.internal.configuration.persistence.listener;
 
 import com.liferay.document.library.internal.configuration.DLSizeLimitConfiguration;
-import com.liferay.document.library.internal.configuration.admin.service.DLSizeLimitManagedServiceFactory;
 import com.liferay.document.library.internal.util.MimeTypeSizeLimitUtil;
 import com.liferay.portal.configuration.persistence.listener.ConfigurationModelListener;
 import com.liferay.portal.configuration.persistence.listener.ConfigurationModelListenerException;
@@ -15,7 +14,6 @@ import com.liferay.portal.kernel.util.ArrayUtil;
 import java.util.Dictionary;
 
 import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Adolfo Pérez
