+++ modules/apps/document-library/document-library-service/src/test/java/com/liferay/document/library/internal/configuration/persistence/listener/DLSizeLimitConfigurationModelListenerTest.java
@@ -5,19 +5,15 @@
 
 package com.liferay.document.library.internal.configuration.persistence.listener;
 
-import com.liferay.document.library.internal.configuration.admin.service.DLSizeLimitManagedServiceFactory;
 import com.liferay.portal.configuration.persistence.listener.ConfigurationModelListenerException;
 import com.liferay.portal.kernel.test.util.RandomTestUtil;
 import com.liferay.portal.kernel.util.HashMapDictionary;
 import com.liferay.portal.kernel.util.HashMapDictionaryBuilder;
 import com.liferay.portal.test.rule.LiferayUnitTestRule;
 
-import org.junit.Before;
 import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.Test;
-
-import org.mockito.Mockito;
 
 /**
  * @author Adolfo Pérez
